# bingle_webserver/spec/server_openapi.yaml
openapi: 3.0.3
info:
  title: Bingle Server API
  version: 0.1.2
  description: |
    Local HTTP server API to allow a browser client to call the Bingle API.
servers:
  - url: http://localhost:12121
    description: Local Bingle server
paths:
  /handleLookup:
    get:
      summary: Lookup an id by handle
      parameters:
        - name: handle
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The associated id
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Handle not found
  /sendMessageToId:
    post:
      summary: Send a message to a user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, message]
              properties:
                userId:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: boolean
  /sendMessageToHandle:
    post:
      summary: Send a message to a handle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle, message]
              properties:
                handle:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: boolean
  /sendMessageToNetwork:
    post:
      summary: Send a message to a network source key and user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [networkSourceKey, userId, message]
              properties:
                networkSourceKey:
                  $ref: '#/components/schemas/NetworkSourceKey'
                userId:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: boolean
  /sendMessageToIdWithResponse:
    post:
      summary: Send a message to a user id and wait for response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, message]
              properties:
                userId:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: The response message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /sendMessageToHandleWithResponse:
    post:
      summary: Send a message to a handle and wait for response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle, message]
              properties:
                handle:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: The response message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /sendMessageToNetworkWithResponse:
    post:
      summary: Send a message to a network source key and user id and wait for response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [networkSourceKey, userId, message]
              properties:
                networkSourceKey:
                  $ref: '#/components/schemas/NetworkSourceKey'
                userId:
                  type: string
                message:
                  $ref: '#/components/schemas/BingleMessage'
      responses:
        '200':
          description: The response message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /queued:
    get:
      summary: Return all received messages queued in the server
      responses:
        '200':
          description: List of queued messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BingleMessage'
  /version:
    get:
      summary: Get current server version information
      responses:
        '200':
          description: Version information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionInfo'
  /getNatType:
    get:
      summary: Get the current detected NAT type
      responses:
        '200':
          description: The current NAT type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NatTypeResponse'
  # Local storage and contacts API (maps BingleLocalApi methods)
  /local/generateKeypair:
    post:
      summary: Generate a new Algorand keypair and set it as current (local API)
      responses:
        '200':
          description: Generated keypair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keypair'
  /local/registerKeypair:
    post:
      summary: Register the current keypair with Bingle using the provided handle (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterKeypairRequest'
      responses:
        '200':
          description: Registration result
          content:
            application/json:
              schema:
                type: boolean
        '400':
          description: Missing configuration or keypair
  /local/addContact:
    post:
      summary: Add a contact to the local store (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContactRequest'
      responses:
        '200':
          description: Added
  /local/blockContact:
    post:
      summary: Block a contact by id (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
      responses:
        '200':
          description: Blocked
        '404':
          description: Contact not found
  /local/removeContact:
    post:
      summary: Remove a contact by id without blocking it (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
      responses:
        '200':
          description: Removed
        '404':
          description: Contact not found
  /local/isBlocked:
    get:
      summary: Check if a contact id is blocked (local API)
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: True if blocked
          content:
            application/json:
              schema:
                type: boolean
  /local/getContacts:
    get:
      summary: Get the list of unblocked contacts (local API)
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
  /local/addMessage:
    post:
      summary: Add a message to the local store (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
      responses:
        '200':
          description: Added
  /local/getMessages:
    get:
      summary: Get the list of stored messages (local API)
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /local/keypairStatus:
    get:
      summary: Check the status of the local keypair (local API)
      responses:
        '200':
          description: Keypair status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeypairStatusResponse'
        '405':
          description: Local API disabled
  /local/save:
    post:
      summary: Save all local state to a JSON file (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: Saved
  /local/load:
    post:
      summary: Load all local state from a JSON file (local API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: Loaded
components:
  schemas:
    VersionInfo:
      type: object
      required: [version, buildTimestamp, buildNumber]
      properties:
        version:
          type: string
        gitSha:
          type: string
          nullable: true
        buildTimestamp:
          type: string
        buildNumber:
          type: string
    NetworkSourceKey:
      type: object
      description: Identifies where to send network traffic (direct or via relay).
      properties:
        inetSocketAddress:
          $ref: '#/components/schemas/InetSocketAddress'
        relayChannel:
          type: integer
          minimum: 0
          maximum: 65535
          description: TURN relay channel number if using a relay.
        relayAddress:
          $ref: '#/components/schemas/InetSocketAddress'
        relayId:
          type: string
          description: Optional relay id (Algorand address).
    InetSocketAddress:
      type: object
      required: [host, port]
      properties:
        host:
          type: string
          description: Hostname or IP address.
        port:
          type: integer
          minimum: 0
          maximum: 65535
          description: UDP port number.
    BingleMessage:
      oneOf:
        - $ref: '../../spec/openapi.yaml#/components/schemas/PlainTextMessage'
        - $ref: '../../spec/openapi.yaml#/components/schemas/PlainTextMessageWithNulls'
        - $ref: '../../spec/openapi.yaml#/components/schemas/TypedMessage'
    Response:
      $ref: '#/components/schemas/BingleMessage'
    # Local API schemas (mirror bingle-local types)
    Keypair:
      type: object
      required: [id, passphrase]
      properties:
        id:
          type: string
          description: Algorand address (public id)
        passphrase:
          type: string
          description: "Algorand mnemonic or base64 seed with b64: prefix"
    ContactSource:
      type: string
      enum: [Manual, Received]
    Contact:
      type: object
      required: [handle, id, fields]
      properties:
        handle:
          type: string
        id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    Message:
      type: object
      required: [senderHandle, recipientHandles, timestamp, text]
      properties:
        senderHandle:
          type: string
        recipientHandles:
          type: array
          items:
            type: string
        timestamp:
          type: integer
          format: int64
        text:
          type: string
    RegisterKeypairRequest:
      type: object
      required: [handle]
      properties:
        handle:
          type: string
          description: The user's unique handle to register on-chain
    AddContactRequest:
      type: object
      required: [handle, id, source]
      properties:
        handle:
          type: string
        id:
          type: string
        source:
          $ref: '#/components/schemas/ContactSource'
    IdRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    AddMessageRequest:
      type: object
      required: [senderHandle, recipientHandles, timestamp, text]
      properties:
        senderHandle:
          type: string
        recipientHandles:
          type: array
          items:
            type: string
        timestamp:
          type: integer
          format: int64
        text:
          type: string
    KeypairStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [None, UNFUNDED, FUNDED, ACTIVE]
          description: |
            None - no keypair exists
            UNFUNDED - keypair exists, no Bingle$ asset, insufficient ALGO
            FUNDED - keypair exists, no Bingle$ asset, sufficient ALGO
            ACTIVE - keypair exists and has Bingle$ asset
        id:
          type: string
          description: Algorand address (present when status is not None)
        handle:
          type: string
          description: Registered handle (present when status is ACTIVE)
        requiredAlgo:
          type: number
          format: double
          description: Required ALGO amount (present when status is UNFUNDED)
    NatTypeResponse:
      type: object
      required: [natType]
      properties:
        natType:
          type: string
          enum: [Unknown, NoConnection, Symmetric, Restricted, FullCone]
          description: |
            Unknown - NAT type has not been determined yet
            NoConnection - no network connection detected
            Symmetric - symmetric NAT (not traversable)
            Restricted - restricted cone NAT (traversable via relay)
            FullCone - full cone NAT (directly reachable)
    PathRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: Filesystem path to a JSON file